requirements.txt
matplotlib
bokeh==2.4.3
cloudpickle
scipy
summa
faiss-cpu
langchain
langchain_openai
langchain_community
langchain_core
langchainhub
openai
instructor
pydantic
feedparser
tiktoken
chromadb
streamlit==1.37.0
streamlit-extras
nltk
cohere
duckduckgo-search
pytextrank
spacy==3.7.5
https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl